+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/managers/scene_entity_cfg.py
@@ -38,8 +38,9 @@ class SceneEntityCfg:
     function as a list of joint indices under :attr:`joint_ids`.
     """
 
-    joint_ids: list[int] | None = None
-    """The indices of the joints from the asset required by the term. Defaults to None.
+    joint_ids: list[int] | slice = slice(None)
+    """The indices of the joints from the asset required by the term. Defaults to slice(None), which means
+    all the joints in the asset (if present).
 
     If :attr:`joint_names` is specified, this is filled in automatically on initialization of the
     manager.
@@ -54,8 +55,9 @@ class SceneEntityCfg:
     function as a list of body indices under :attr:`body_ids`.
     """
 
-    body_ids: list[int] | None = None
-    """The indices of the bodies from the asset required by the term. Defaults to None.
+    body_ids: list[int] | slice = slice(None)
+    """The indices of the bodies from the asset required by the term. Defaults to slice(None), which means
+    all the bodies in the asset.
 
     If :attr:`body_names` is specified, this is filled in automatically on initialization of the
     manager.
@@ -79,11 +81,12 @@ class SceneEntityCfg:
         # check if the entity is valid
         if self.name not in scene.keys():
             raise ValueError(f"The scene entity '{self.name}' does not exist. Available entities: {scene.keys()}.")
+
         # convert joint names to indices based on regex
-        if self.joint_names is not None or self.joint_ids is not None:
+        if self.joint_names is not None or self.joint_ids != slice(None):
             entity: Articulation = scene[self.name]
-            # -- if both are not None, check if they are valid
-            if self.joint_names is not None and self.joint_ids is not None:
+            # -- if both are not their default values, check if they are valid
+            if self.joint_names is not None and self.joint_ids != slice(None):
                 if isinstance(self.joint_names, str):
                     self.joint_names = [self.joint_names]
                 if isinstance(self.joint_ids, int):
@@ -102,16 +105,20 @@ class SceneEntityCfg:
                 if isinstance(self.joint_names, str):
                     self.joint_names = [self.joint_names]
                 self.joint_ids, _ = entity.find_joints(self.joint_names)
+                # performance optimization (slice offers faster indexing than list of indices)
+                if len(self.joint_ids) == entity.num_joints:
+                    self.joint_ids = slice(None)
             # -- from joint indices to joint names
-            elif self.joint_ids is not None:
+            elif self.joint_ids != slice(None):
                 if isinstance(self.joint_ids, int):
                     self.joint_ids = [self.joint_ids]
                 self.joint_names = [entity.joint_names[i] for i in self.joint_ids]
+
         # convert body names to indices based on regex
-        if self.body_names is not None or self.body_ids is not None:
+        if self.body_names is not None or self.body_ids != slice(None):
             entity: RigidObject = scene[self.name]
-            # -- if both are not None, check if they are valid
-            if self.body_names is not None and self.body_ids is not None:
+            # -- if both are not their default values, check if they are valid
+            if self.body_names is not None and self.body_ids != slice(None):
                 if isinstance(self.body_names, str):
                     self.body_names = [self.body_names]
                 if isinstance(self.body_ids, int):
@@ -130,8 +137,11 @@ class SceneEntityCfg:
                 if isinstance(self.body_names, str):
                     self.body_names = [self.body_names]
                 self.body_ids, _ = entity.find_bodies(self.body_names)
+                # performance optimization (slice offers faster indexing than list of indices)
+                if len(self.body_ids) == entity.num_bodies:
+                    self.body_ids = slice(None)
             # -- from body indices to body names
-            elif self.body_ids is not None:
+            elif self.body_ids != slice(None):
                 if isinstance(self.body_ids, int):
                     self.body_ids = [self.body_ids]
                 self.body_names = [entity.body_names[i] for i in self.body_ids]